Refactored code to game core
