+++ src/backend/gameLogic/game_logic.py
@@ -1,0 +1,3 @@
+class GameLogic:
+    def __init__(self) -> None:
+        pass
